abandon mutex (posix_ringbufr)
The popInquire method was too conservative when there was thread
contention. No data corruption, but popInquire returned zero length
when the buffer was not empty. So I removed support for thread safety.
Users will have to protect all four RingbufR methods with a single
mutex. This is now demonstrated in program testring.